+++ src/onnx/parse_range.cpp
@@ -34,7 +34,7 @@ struct parse_range : op_parser<parse_range>
             auto limit_val = limit.front();
             auto delta_val = delta.front();
 
-            size_t num_elements = static_cast<size_t>(
+            int num_elements = static_cast<int>(
                 ceil(static_cast<double>(limit_val - start_val) / static_cast<double>(delta_val)));
 
             assert(num_elements > 0);
