+++ test/verify/test_multinomial.cpp
@@ -10,8 +10,8 @@ struct test_multinomial : verify_program<test_multinomial>
     {
         migraphx::program p;
         auto* mm           = p.get_main_module();
-        size_t sample_size = 10;
-        size_t batch_size  = 2;
+        int sample_size = 10;
+        int batch_size  = 2;
         float seed         = 0.0f;
         std::mt19937 gen(seed);
         std::uniform_real_distribution<> dis(0.0, 1.0);
